refactor external api into folder
