+++ app/src/forms/form/index.js
@@ -1,6 +1,9 @@
 const routes = require('./routes');
 const setupMount = require('../common/utils').setupMount;
 
+const externalApiRoutes = require('./externalApi/routes');
+
 module.exports.mount = (app) => {
-  return setupMount('forms', app, routes);
+  const p = setupMount('forms', app, [routes, externalApiRoutes]);
+  return p;
 };
