+++ app/src/forms/form/routes.js
@@ -171,28 +171,4 @@ routes.put('/:formId/subscriptions', hasFormPermissions([P.FORM_READ, P.FORM_UPD
   await controller.createOrUpdateSubscriptionDetails(req, res, next);
 });
 
-routes.get('/:formId/externalAPIs', hasFormPermissions([P.FORM_READ, P.FORM_UPDATE]), async (req, res, next) => {
-  await controller.listExternalAPIs(req, res, next);
-});
-
-routes.post('/:formId/externalAPIs', hasFormPermissions([P.FORM_READ, P.FORM_UPDATE]), async (req, res, next) => {
-  await controller.createExternalAPI(req, res, next);
-});
-
-routes.get('/:formId/externalAPIs/algorithms', hasFormPermissions([P.FORM_READ, P.FORM_UPDATE]), async (req, res, next) => {
-  await controller.listExternalAPIAlgorithms(req, res, next);
-});
-
-routes.get('/:formId/externalAPIs/statusCodes', hasFormPermissions([P.FORM_READ, P.FORM_UPDATE]), async (req, res, next) => {
-  await controller.listExternalAPIStatusCodes(req, res, next);
-});
-
-routes.put('/:formId/externalAPIs/:externalAPIId', hasFormPermissions([P.FORM_READ, P.FORM_UPDATE]), async (req, res, next) => {
-  await controller.updateExternalAPI(req, res, next);
-});
-
-routes.delete('/:formId/externalAPIs/:externalAPIId', hasFormPermissions([P.FORM_READ, P.FORM_UPDATE]), async (req, res, next) => {
-  await controller.deleteExternalAPI(req, res, next);
-});
-
 module.exports = routes;
