+++ app/tests/unit/forms/form/externalApi/service.spec.js
@@ -1,0 +1,197 @@
+const { MockModel, MockTransaction } = require('../../../../common/dbHelper');
+
+const { v4: uuidv4 } = require('uuid');
+
+const { ExternalAPIStatuses } = require('../../../../../src/forms/common/constants');
+const service = require('../../../../../src/forms/form/externalApi/service');
+const { ENCRYPTION_ALGORITHMS } = require('../../../../../src/components/encryptionService');
+
+jest.mock('../../../../../src/forms/common/models/tables/externalAPI', () => MockModel);
+
+beforeEach(() => {
+  MockModel.mockReset();
+  MockTransaction.mockReset();
+});
+
+afterEach(() => {
+  jest.restoreAllMocks();
+});
+
+describe('validateExternalAPI', () => {
+  let validData = null;
+  beforeEach(() => {
+    validData = {
+      id: uuidv4(),
+      formId: uuidv4(),
+      name: 'test_api',
+      endpointUrl: 'http://external.api/',
+      sendApiKey: true,
+      apiKeyHeader: 'X-API-KEY',
+      apiKey: 'my-api-key',
+      sendUserToken: true,
+      userTokenHeader: 'Authorization',
+      userTokenBearer: true,
+      sendUserInfo: true,
+      userInfoHeader: 'X-API-USER',
+      userInfoEncrypted: true,
+      userInfoEncryptionKey: '0489aa2a7882dc53be7c76db43be1800e56627c31a88a0011d85ccc255b79d00',
+      userInfoEncryptionAlgo: ENCRYPTION_ALGORITHMS.AES_256_GCM,
+      code: ExternalAPIStatuses.SUBMITTED,
+    };
+  });
+  it('should not throw errors with valid data', () => {
+    service.validateExternalAPI(validData);
+  });
+
+  it('should throw 422 with no data', () => {
+    expect(() => service.validateExternalAPI(undefined)).toThrow();
+  });
+
+  it('should throw 422 when userInfo encryption options are invalid', () => {
+    validData.userInfoEncryptionKey = null;
+    expect(() => service.validateExternalAPI(undefined)).toThrow();
+  });
+
+  it('should throw 422 when userInfoEncryptionAlgo is invalid', () => {
+    validData.userInfoEncryptionAlgo = 'not valid!';
+    expect(() => service.validateExternalAPI(undefined)).toThrow();
+  });
+
+  it('should throw 422 when sendApiKey is true with no header', () => {
+    validData.apiKeyHeader = null;
+    expect(() => service.validateExternalAPI(undefined)).toThrow();
+  });
+
+  it('should throw 422 when sendApiKey is true with no key', () => {
+    validData.apiKey = null;
+    expect(() => service.validateExternalAPI(undefined)).toThrow();
+  });
+
+  it('should throw 422 when sendUserInfo (encrypted) is true with no header', () => {
+    validData.userInfoHeader = null;
+    expect(() => service.validateExternalAPI(undefined)).toThrow();
+  });
+
+  it('should throw 422 when sendUserInfo (encrypted) is true with no header', () => {
+    validData.userInfoHeader = null;
+    expect(() => service.validateExternalAPI(undefined)).toThrow();
+  });
+});
+
+describe('createExternalAPI', () => {
+  const user = { usernameIdp: 'username' };
+  let validData = null;
+
+  beforeEach(() => {
+    MockModel.mockReset();
+    MockTransaction.mockReset();
+    validData = {
+      id: uuidv4(),
+      formId: uuidv4(),
+      name: 'test_api',
+      endpointUrl: 'http://external.api/',
+      sendApiKey: true,
+      apiKeyHeader: 'X-API-KEY',
+      apiKey: 'my-api-key',
+      sendUserToken: true,
+      userTokenHeader: 'Authorization',
+      userTokenBearer: true,
+      sendUserInfo: true,
+      userInfoHeader: 'X-API-USER',
+      userInfoEncrypted: true,
+      userInfoEncryptionKey: '0489aa2a7882dc53be7c76db43be1800e56627c31a88a0011d85ccc255b79d00',
+      userInfoEncryptionAlgo: ENCRYPTION_ALGORITHMS.AES_256_GCM,
+      code: ExternalAPIStatuses.SUBMITTED,
+    };
+  });
+
+  afterEach(() => {
+    jest.restoreAllMocks();
+  });
+
+  it('should insert valid data', async () => {
+    validData.id = null;
+    validData.code = null;
+    await service.createExternalAPI(validData.formId, validData, user);
+    expect(MockModel.insert).toBeCalledTimes(1);
+    expect(MockModel.insert).toBeCalledWith({
+      createdBy: user.usernameIdp,
+      code: ExternalAPIStatuses.SUBMITTED,
+      ...validData,
+    });
+    expect(MockTransaction.commit).toBeCalledTimes(1);
+  });
+
+  it('should rollback on error', async () => {
+    MockModel.insert = jest.fn().mockRejectedValueOnce(new Error('SQL Error'));
+
+    await expect(service.createExternalAPI(validData.formId, validData, user)).rejects.toThrow();
+
+    expect(MockTransaction.rollback).toBeCalledTimes(1);
+  });
+});
+
+describe('updateExternalAPI', () => {
+  const user = { usernameIdp: 'username' };
+  let validData = null;
+
+  beforeEach(() => {
+    MockModel.mockReset();
+    MockTransaction.mockReset();
+    validData = {
+      id: uuidv4(),
+      formId: uuidv4(),
+      name: 'test_api',
+      endpointUrl: 'http://external.api/',
+      sendApiKey: true,
+      apiKeyHeader: 'X-API-KEY',
+      apiKey: 'my-api-key',
+      sendUserToken: true,
+      userTokenHeader: 'Authorization',
+      userTokenBearer: true,
+      sendUserInfo: true,
+      userInfoHeader: 'X-API-USER',
+      userInfoEncrypted: true,
+      userInfoEncryptionKey: '0489aa2a7882dc53be7c76db43be1800e56627c31a88a0011d85ccc255b79d00',
+      userInfoEncryptionAlgo: ENCRYPTION_ALGORITHMS.AES_256_GCM,
+      code: ExternalAPIStatuses.SUBMITTED,
+    };
+  });
+
+  afterEach(() => {
+    jest.restoreAllMocks();
+  });
+
+  it('should update valid data', async () => {
+    MockModel.throwIfNotFound = jest.fn().mockResolvedValueOnce(Object.assign({}, validData));
+
+    // we do not update (status) code - must stay SUBMITTED
+    validData.code = ExternalAPIStatuses.APPROVED;
+
+    await service.updateExternalAPI(validData.formId, validData.id, validData, user);
+    expect(MockModel.update).toBeCalledTimes(1);
+    expect(MockModel.update).toBeCalledWith({
+      updatedBy: user.usernameIdp,
+      code: ExternalAPIStatuses.SUBMITTED,
+      ...validData,
+    });
+    expect(MockTransaction.commit).toBeCalledTimes(1);
+  });
+
+  it('should not commit when not found', async () => {
+    MockModel.throwIfNotFound = jest.fn().mockRejectedValueOnce(new Error('SQL Error'));
+
+    await expect(service.updateExternalAPI(validData.formId, validData.id, validData, user)).rejects.toThrow();
+    // shouldn't start the transaction
+    expect(MockModel.startTransaction).toBeCalledTimes(0);
+    expect(MockTransaction.commit).toBeCalledTimes(0);
+  });
+
+  it('should rollback on error', async () => {
+    MockModel.update = jest.fn().mockRejectedValueOnce(new Error('SQL Error'));
+
+    await expect(service.updateExternalAPI(validData.formId, validData.id, validData, user)).rejects.toThrow();
+
+    expect(MockTransaction.rollback).toBeCalledTimes(1);
+  });
+});
